Changed button styling and indicate which page is active
