+++ frontend/src/components/navbar/navbar.tsx
@@ -1,24 +1,36 @@
 import './Navbar.css';
 
-import Button from '../assets/Button';
+import Button from '../assets/button/Button';
 
 interface NavbarProps {
     setCurrentPage: (page: string) => void;
+    currentPage: string;
 }
 
-export default function Navbar({ setCurrentPage }: NavbarProps) {
+export default function Navbar({ setCurrentPage, currentPage }: NavbarProps) {
+    const isActive = (page: string) => currentPage === page ? 'active' : '';
+
     return (
         <nav className="navbar">
             <div className="navbar-left">
-                <a href="/" className="logo">Page Title</a>
+                <a href="/" className="logo">Logo</a>
             </div>
             <div className="navbar-center">
                 
             </div>
             <div className="navbar-right">
                 <ul className="nav-buttons">
-                    <Button buttonText={'HOME'} onClick={() => setCurrentPage('home')}></Button>
-                    <Button buttonText={'PROFILE'} onClick={() => setCurrentPage('profile')}></Button>
+                    <Button 
+                    buttonStyle={`nav-button ${isActive('home')}`}
+                    buttonText={'HOME'} 
+                    onClick={() => setCurrentPage('home')}
+                    ></Button>
+
+                    <Button 
+                    buttonStyle={`nav-button ${isActive('profile')}`}
+                    buttonText={'PROFILE'} 
+                    onClick={() => setCurrentPage('profile')}
+                    ></Button>
                 </ul>
             </div>
         </nav>
